Just query the db directly. test rows code.

+++ n1mmlistener/ContactDbRow.cs
@@ -40,6 +40,6 @@ namespace n1mmlistener
         /// <summary>
         /// MHz
         /// </summary>
-        public int Band { get; set; }
+        public double Band { get; set; }
     }
 }
